F4fixauth

+++ src/main/java/com/santatecla/G1/author/AuthorRestController.java
@@ -30,12 +30,16 @@ import com.santatecla.G1.user.UserComponent;
 @RestController
 @RequestMapping("/api")
 public class AuthorRestController {
-	interface AuthorDetailView extends Author.NameView, Author.BasicView, Author.BooksView,Book.NameView, Book.BasicView, Book.ThemeView, Theme.BasicView, Book.CitationsView, Citation.BasicView {}
-	interface AuthorBasicView extends Author.NameView, Author.BasicView {}
-	
+	interface AuthorDetailView extends Author.NameView, Author.BasicView, Author.BooksView, Book.NameView,
+			Book.BasicView, Book.ThemeView, Theme.BasicView, Book.CitationsView, Citation.BasicView {
+	}
+
+	interface AuthorBasicView extends Author.NameView, Author.BasicView {
+	}
+
 	@Autowired
 	private UserComponent userComponent;
-	
+
 	@Autowired
 	private AuthorService authorService;
 
@@ -65,28 +69,29 @@ public class AuthorRestController {
 			}
 		}
 		MappingJacksonValue result = new MappingJacksonValue(authors);
-		if(authors!=null) {
-			if(userComponent.isLoggedUser())
+		if (authors != null) {
+			if (userComponent.isLoggedUser())
 				result.setSerializationView(AuthorBasicView.class);
 			else
 				result.setSerializationView(Author.NameView.class);
 			return result;
-		}
-		else return null;
+		} else
+			return null;
 	}
-	
 
 	@JsonView(AuthorDetailView.class)
 	@RequestMapping(value = "/authors/{id}", method = GET)
 	public ResponseEntity<Author> getAuthor(@PathVariable long id) {
-		Author a = authorService.findById(id);
-		if (a != null)
-			return new ResponseEntity<>(a, HttpStatus.OK);
-		else
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		if (userComponent.isLoggedUser()) {
+			Author a = authorService.findById(id);
+			if (a != null)
+				return new ResponseEntity<>(a, HttpStatus.OK);
+			else
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		} else
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
-	
-	
+
 	@JsonView(AuthorDetailView.class)
 	@RequestMapping(value = "/authors", method = RequestMethod.POST)
 	public ResponseEntity<Author> author(@RequestBody Author author) {
@@ -94,10 +99,10 @@ public class AuthorRestController {
 			if (author.getBooks() != null) {
 				authorService.save(author);
 				ArrayList<Book> books = new ArrayList<>();
-				Book bookAux= new Book();
+				Book bookAux = new Book();
 				for (Book book : author.getBooks()) {
 					books.add(bookService.findById(book.getId()));
-					bookAux=bookService.findById(book.getId());
+					bookAux = bookService.findById(book.getId());
 					bookAux.setAuthor(author);
 					bookService.save(bookAux);
 				}
@@ -112,57 +117,63 @@ public class AuthorRestController {
 	@JsonView(AuthorDetailView.class)
 	@RequestMapping(value = "/authors/{id}", method = PATCH)
 	public ResponseEntity<Author> updateAuthor(@RequestBody Author newAuthor, @PathVariable long id) {
-		Author oldAuthor = authorService.findById(id);
-		if (oldAuthor != null) {
-			oldAuthor.update(newAuthor);
-			authorService.save(oldAuthor);
-			return new ResponseEntity<>(oldAuthor, HttpStatus.OK);
-		} else {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
+		if (userComponent.isLoggedUser()) {
+			Author oldAuthor = authorService.findById(id);
+			if (oldAuthor != null) {
+				oldAuthor.update(newAuthor);
+				authorService.save(oldAuthor);
+				return new ResponseEntity<>(oldAuthor, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+		} else
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 
 	@RequestMapping(value = "/authors/{id}", method = DELETE)
 	public ResponseEntity<Author> deleteAuthor(@PathVariable long id) {
-		Author author = authorService.findById(id);
-		if (author != null) {
-			if (author.getBooks()!=null) {
-				author.setBooks(null);
-				authorService.save(author);
-				authorService.deleteById(id);
-				return new ResponseEntity<>(author, HttpStatus.OK);
-			}
-			else {
-				return new ResponseEntity<>(HttpStatus.IM_USED);
-			}
-			
-		} else {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
+		if (userComponent.isLoggedUser()) {
+			Author author = authorService.findById(id);
+			if (author != null) {
+				if (author.getBooks() != null) {
+					author.setBooks(null);
+					authorService.save(author);
+					authorService.deleteById(id);
+					return new ResponseEntity<>(author, HttpStatus.OK);
+				} else {
+					return new ResponseEntity<>(HttpStatus.IM_USED);
+				}
 
+			} else {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+		} else
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
-	
-	
-	
-	
-	// ----------------------------- METHODS WITH UPLOAD IMAGES -------------------------------------------------
-	
-	
-	
+
+	// ----------------------------- METHODS WITH UPLOAD IMAGES
+	// -------------------------------------------------
+
 	@JsonView(AuthorDetailView.class)
 	@RequestMapping(value = "/authors/{id}/image", method = PATCH)
-	public ResponseEntity<Author> updateAuthorImage(Model model, @PathVariable long id, @RequestParam(value="file")MultipartFile file) {
-		Author oldAuthor = authorService.findById(id);
-		if (oldAuthor != null) {
-			if ((file != null) && (!file.isEmpty())) {
-				int imgId = com.santatecla.G1.image.ImageManagerController.getNextId();
-				com.santatecla.G1.image.ImageManagerController.handleFileUpload(model, file, imgId);
-				oldAuthor.setImgId(imgId);
+	public ResponseEntity<Author> updateAuthorImage(Model model, @PathVariable long id,
+			@RequestParam(value = "file") MultipartFile file) {
+		if (userComponent.isLoggedUser()) {
+			Author oldAuthor = authorService.findById(id);
+			if (oldAuthor != null) {
+				if ((file != null) && (!file.isEmpty())) {
+					int imgId = com.santatecla.G1.image.ImageManagerController.getNextId();
+					com.santatecla.G1.image.ImageManagerController.handleFileUpload(model, file, imgId);
+					oldAuthor.setImgId(imgId);
+				}
+				authorService.save(oldAuthor);
+				return new ResponseEntity<>(oldAuthor, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 			}
-			authorService.save(oldAuthor);
-			return new ResponseEntity<>(oldAuthor, HttpStatus.OK);
-		} else {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
+		} else
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 }
+
+
